// frontend/src/generated/http/client/models/MediaItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MediaItem = {
    captured_at_ms?: number | null;
    content_type: string;
    description?: string | null;
    fps?: number | null;
    height?: number | null;
    imu_data_file_name?: string | null;
    imu_data_samples?: number | null;
    kind?: string | null;
    label_attached?: boolean;
    label_file_name?: string | null;
    model_id?: string | null;
    model_input_resolution?: string | null;
    model_tensor_spec?: string | null;
    name: string;
    size_bytes: number;
    stream_id?: string | null;
    tags?: Array<string>;
    video_codec?: string | null;
    width?: number | null;
};