// frontend/src/generated/http/client/models/UpdateMetadataRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateMetadataRequest = {
    description?: string | null;
    model_input_resolution?: string | null;
    model_tensor_spec?: string | null;
    name?: string | null;
    tags?: any[] | null;
};